Moved input into proper folder.
