+++ src/openssd-dashboard/popups/about-popup/about-popup.component.ts
@@ -18,7 +18,7 @@
 
 import { ChangeDetectionStrategy, Component, Inject } from '@angular/core';
 import { PopupReference } from '../../shared/components/popup/popup.service';
-import { InputComponent } from '../../shared/components/search/input.component';
+import { InputComponent } from '../../shared/components/input/input.component';
 import { ButtonComponent } from '../../shared/components/button/button.component';
 import { FormsModule } from '@angular/forms';
