+++ src/openssd-dashboard/shared/components/input/input.component.scss
